chore: split networks.py
